+++ components/AccountSection/OrderHistorySection.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const OrderHistorySection = (props: Props) => {
+  return (
+    <div>OrderHistorySection</div>
+  )
+}
+
+export default OrderHistorySection
